+++ mvc/src/main/java/com/interface21/webmvc/servlet/mvc/tobe/adapter/HandlerAdapterRegistry.java
@@ -1,4 +1,4 @@
-package com.interface21.webmvc.servlet.mvc.tobe;
+package com.interface21.webmvc.servlet.mvc.tobe.adapter;
 
 import java.util.ArrayList;
 import java.util.List;
